requirements.txt
asgiref==3.7.2
dj-rest-auth==5.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.2
django-guardian==2.4.0
djangorestframework==3.14.0
Markdown==3.4.4
psycopg2-binary==2.9.7
pytz==2023.3.post1
sqlparse==0.4.4
